수정: import 오류 및 필요 없는 기능 삭제
